Nearly completely immutable model introduced.
Player moves work
The ghosts move once but are then dangling, due to the caching that the task does
of the old ghost objects.

The animations are however still mutable and will need either internal synchronisation
or their own immutable model.

Due to the new design of how we manage entities collision detection is
more straight forward. Since all type information is kept there is no need
to regain this information through type checks or reflection as was done in the old
player collisions and CollisionInteractionMaps.

All tests are horribly broken.