Fix navigation behaviour for ghosts
The occupants list in the square class was not used and is now removed.
However the Navigation class relied on this information to find units.

To fix this problem the navigation methods and ghost move methods, now
accept a Entities parameter to provide the needed information, thus
depending on it explicitly rather than implicitly.